+++ source/write/worksheet.js
@@ -11,7 +11,7 @@ import generateColumnsDescription from './columns'
 const WORKSHEET_TEMPLATE = `<?xml version="1.0" ?>
 <worksheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:mx="http://schemas.microsoft.com/office/mac/excel/2008/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:x14="http://schemas.microsoft.com/office/spreadsheetml/2009/9/main" xmlns:x14ac="http://schemas.microsoft.com/office/spreadsheetml/2009/9/ac" xmlns:xm="http://schemas.microsoft.com/office/excel/2006/main">{columnsDescription}<sheetData>{data}</sheetData></worksheet>`;
 
-export default function generateWorksheet(data, { schema, columns, getStyle, getSharedString }) {
+export default function generateWorksheet(data, { schema, columns, getStyle, getSharedString, customFont }) {
 	if (schema) {
 		if (!Array.isArray(data)) {
 			throw new TypeError('Expected an array of objects')
@@ -27,6 +27,6 @@ export default function generateWorksheet(data, { schema, columns, getStyle, get
 		}
 	}
   return WORKSHEET_TEMPLATE
-  	.replace('{data}', generateRows(data, { schema, getStyle, getSharedString }))
+  	.replace('{data}', generateRows(data, { schema, getStyle, getSharedString, customFont }))
   	.replace('{columnsDescription}', generateColumnsDescription({ schema, columns }))
 }